backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
python-dotenv
mysql-connector-python
Pillow
channels
channels_redis
daphne